make it so you can import fake data functions without wiping the database
